solution/question4/requirements.txt
aws-cdk-lib==2.130.0
constructs>=10.0.0,<11.0.0
boto3
python-dotenv
